studio: properly load showcases on start
