+++ packages/legend-application-studio/src/components/editor/EditorStoreProvider.tsx
@@ -36,7 +36,6 @@ export const EditorStoreProvider: React.FC<{
         applicationStore,
         baseStore.sdlcServerClient,
         baseStore.depotServerClient,
-        baseStore.showcaseManagerState,
       ),
   );
   return (
