+++ packages/legend-application-studio/src/components/editor/__test-utils__/EditorComponentTestUtils.tsx
@@ -65,10 +65,7 @@ import {
   createMemoryHistory,
   TEST__BrowserEnvironmentProvider,
 } from '@finos/legend-application/test';
-import {
-  ShowcaseManagerState,
-  type LegendStudioApplicationStore,
-} from '../../../stores/LegendStudioBaseStore.js';
+import { type LegendStudioApplicationStore } from '../../../stores/LegendStudioBaseStore.js';
 import { TEST__getLegendStudioApplicationConfig } from '../../../stores/__test-utils__/LegendStudioApplicationTestUtils.js';
 import { Route } from '@finos/legend-application/browser';
 
@@ -161,7 +158,6 @@ export const TEST__provideMockedEditorStore = (customization?: {
       new DepotServerClient({
         serverUrl: applicationStore.config.depotServerUrl,
       }),
-      new ShowcaseManagerState(applicationStore),
     );
   const MOCK__EditorStoreProvider = require('../EditorStoreProvider.js'); // eslint-disable-line @typescript-eslint/no-unsafe-assignment
   MOCK__EditorStoreProvider.useEditorStore = createMock();
