+++ packages/legend-application-studio/src/stores/editor/__test-utils__/EditorStoreTestUtils.ts
@@ -20,7 +20,6 @@ import { SDLCServerClient } from '@finos/legend-server-sdlc';
 import { DepotServerClient } from '@finos/legend-server-depot';
 import { ApplicationStore } from '@finos/legend-application';
 import { TEST__getLegendStudioApplicationConfig } from '../../__test-utils__/LegendStudioApplicationTestUtils.js';
-import { ShowcaseManagerState } from '../../LegendStudioBaseStore.js';
 
 export const TEST__getTestEditorStore = (
   pluginManager = LegendStudioPluginManager.create(),
@@ -40,6 +39,5 @@ export const TEST__getTestEditorStore = (
     new DepotServerClient({
       serverUrl: applicationStore.config.depotServerUrl,
     }),
-    new ShowcaseManagerState(applicationStore),
   );
 };
