+++ packages/legend-extension-dsl-service/src/stores/studio/ServiceQueryEditorStore.ts
@@ -26,7 +26,6 @@ import {
   MINIMUM_SERVICE_OWNERS,
   generateServiceManagementUrl,
   pureExecution_setFunction,
-  ShowcaseManagerState,
 } from '@finos/legend-application-studio';
 import {
   type Service,
@@ -100,12 +99,7 @@ export abstract class ServiceQueryEditorStore extends EditorStore {
     sdlcServerClient: SDLCServerClient,
     depotServerClient: DepotServerClient,
   ) {
-    super(
-      applicationStore,
-      sdlcServerClient,
-      depotServerClient,
-      new ShowcaseManagerState(applicationStore),
-    );
+    super(applicationStore, sdlcServerClient, depotServerClient);
 
     makeObservable(this, {
       queryBuilderState: observable,
